+++ apiv3/api-tester/case/user.js
@@ -1,16 +1,17 @@
 var async = require("async")
 var common = require("../common");
-var auth = require("./auth");
 var hippie = common.hippie;
 var swaggerHippie = common.swaggerHippie;
 var expect = common.expect;
+
+var authApi = require("../api/auth");
 
 module.exports = function(finalDone) {
     //newNodeId = "2e9790e8c4334f65b332a65f9553cca3";
     //newClusterId = 96;
     async.series([
-        function(callback) { auth.authPost(callback) },
-        function(callback) { userGet(callback) },
+        function(callback) { authApi.authPost(callback) },
+        function(callback) { userGet(callback) }
     ], finalDone);
 };
